Added a starting scene with basic movement and some ground work code.
Did a bit of sprite animation between the skeleton and the necromancer. Both of which are pretty rough but will work for now. Only have them facing one way or the other, later on we could add in a walking up and walking down animation, but that'll go later on.

Added some basic Post Processing to make the game look good and added Unity 2D lights. Even adding a normal map to the ground texture for some extra realism.

Movement and aiming are split up among a bunch of scripts that ive documented as well as I can. Honestly not sure how much I like this initial character setup. The Unity input system side of it is good, but this new style of compartmentalizing all the code of the character into a bunch of small scripts works well now, but could eventually get just as cluttered as a mega character script.

Equipment framework is mostly a place holder for now. We can make adjustments to the setup of these scripts once we find out if these kits are just going to be weapon types or if the kit will have other passives/abilities. Like a +2 dmg warrior kit.

Also have the base skeleton loop just playing atm till I bother to code out the minions. Additionally have a visual mock up of a little panel that can popup when the necromancer gets close to a raise-able corpse.