Generation works for sure this time
Added the door carving and colors for different room types. Everything shown is just for visual debugging. Generated a bunch of rooms and couldn't find anything that didn't work. Also added a seed with a bool so that if while generating we do find one that is broken I can look at it and figure out why.