Creating Interfaces for combat system
I've gotta change the mechanics for existing kits with this new interpretation for combat, but other than that things are looking pretty good. I've cleaned up the code a bit to make things even more modular and expandable. If I managed to set this up correctly, we'll be able to greatly increase the AI's versatility.